Use collectorType to distinguish between hosted/installable collectors.
