+++ src/main/java/com/sumologic/client/collectors/model/HostedCollector.java
@@ -1,0 +1,10 @@
+package com.sumologic.client.collectors.model;
+
+/**
+ * A hosted collector in the Sumo Logic system.
+ *
+ * @author Jeffrey Wang
+ * @version 1.1
+ */
+public class HostedCollector extends Collector {
+}
